codex-ask round-2 fixes (all three claims verified in source): T80 e3 fusion - clone folded into the kernel (every element written once, zero-KV constants + untouched copies, (2+z)D->(2-z)D traffic) + lse head mask (non-pow2 head rows could be overrun, platform shapes passed pow2 so latent); T77 prefix+offset promoted to int64 before store (i32 overflow near 2^31, both paths) + boundary tests; T79 splits denominator 128->256 aligned with BLOCK_P (half the splits were idle)

+++ src/flaggems_sglang/ops/compute_position.py
@@ -38,7 +38,7 @@ def _compute_position(
         offset = tl.arange(0, BLOCK_TOKENS) + i * BLOCK_TOKENS
         tl.store(
             positions + cumsum_start + offset,
-            prefix_len + offset,
+            prefix_len.to(tl.int64) + offset,
             mask=offset < seq_len,
         )
     tl.store(starts + pid, cumsum_start)
@@ -75,7 +75,7 @@ def _compute_position_striped(
             token_offsets = offsets + tile * BLOCK_TOKENS
             tl.store(
                 positions + cumsum_start + token_offsets,
-                prefix_len + token_offsets,
+                prefix_len.to(tl.int64) + token_offsets,
                 mask=token_offsets < seq_len,
             )
         cumsum_start += seq_len
